requirements.txt
Django>=3.2.0
opencv-python>=4.5.0
numpy>=1.19.0
torch>=1.9.0
torchvision>=0.10.0
facenet-pytorch>=2.5.0
scikit-learn>=0.24.0
Pillow>=8.0.0
django-crispy-forms>=1.12.0
crispy-bootstrap4>=1.0.0
python-dotenv>=0.19.0
face-recognition==1.3.0
pandas==2.1.4
xlsxwriter==3.1.9 